backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
web3
tweepy
python-telegram-bot
APScheduler
redis
